+++ src/com/intellij/idea/plugin/hybris/impex/psi/references/ImpExValueTSEnumReference.kt
@@ -63,10 +63,6 @@ abstract class ImpExValueTSEnumReference(
         ?.toTypedArray()
         ?: emptyArray()
 
-//    override fun resolve(): PsiElement? = multiResolve(false)
-//        .lastOrNull()
-//        ?.element
-
     override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {
         val indicator = ProgressManager.getInstance().progressIndicator
         if (indicator != null && indicator.isCanceled) return ResolveResult.EMPTY_ARRAY